fix bug and refactor crc16 to HashSlotUtils
