+++ libs/cluster/Server/ClusterManagerSlotState.cs
@@ -463,7 +463,7 @@ namespace Garnet.cluster
             while (iter.GetNext(out _))
             {
                 ref SpanByte key = ref iter.GetKey();
-                var s = NumUtils.HashSlot(key.ToPointer(), key.Length);
+                var s = HashSlotUtils.HashSlot(key.ToPointer(), key.Length);
                 if (slots.Contains(s))
                     _ = BasicGarnetApi.DELETE(ref key, StoreType.Main);
             }
@@ -481,7 +481,7 @@ namespace Garnet.cluster
             {
                 ref var key = ref iterObject.GetKey();
                 ref var value = ref iterObject.GetValue();
-                var s = NumUtils.HashSlot(key);
+                var s = HashSlotUtils.HashSlot(key);
                 if (slots.Contains(s))
                     _ = BasicGarnetApi.DELETE(key, StoreType.Object);
             }
