+++ libs/cluster/Session/ClusterKeyIterationFunctions.cs
@@ -23,7 +23,7 @@ namespace Garnet.cluster
                 public bool SingleReader(ref SpanByte key, ref SpanByte value, RecordMetadata recordMetadata, long numberOfRecords, out CursorRecordResult cursorRecordResult)
                 {
                     cursorRecordResult = CursorRecordResult.Accept; // default; not used here
-                    if (NumUtils.HashSlot(key.ToPointer(), key.LengthWithoutMetadata) == slot && !Expired(ref value))
+                    if (HashSlotUtils.HashSlot(key.ToPointer(), key.LengthWithoutMetadata) == slot && !Expired(ref value))
                         keyCount++;
                     return true;
                 }
@@ -46,7 +46,7 @@ namespace Garnet.cluster
                     cursorRecordResult = CursorRecordResult.Accept; // default; not used here , out CursorRecordResult cursorRecordResult
                     fixed (byte* keyPtr = key)
                     {
-                        if (NumUtils.HashSlot(keyPtr, key.Length) == slot && !Expired(ref value))
+                        if (HashSlotUtils.HashSlot(keyPtr, key.Length) == slot && !Expired(ref value))
                             keyCount++;
                     }
                     return true;
@@ -73,7 +73,7 @@ namespace Garnet.cluster
                 public bool SingleReader(ref SpanByte key, ref SpanByte value, RecordMetadata recordMetadata, long numberOfRecords, out CursorRecordResult cursorRecordResult)
                 {
                     cursorRecordResult = CursorRecordResult.Accept; // default; not used here, out CursorRecordResult cursorRecordResult
-                    if (NumUtils.HashSlot(key.ToPointer(), key.LengthWithoutMetadata) == slot && !Expired(ref value))
+                    if (HashSlotUtils.HashSlot(key.ToPointer(), key.LengthWithoutMetadata) == slot && !Expired(ref value))
                         keys.Add(key.ToByteArray());
                     return keys.Count < maxKeyCount;
                 }
@@ -100,7 +100,7 @@ namespace Garnet.cluster
                     cursorRecordResult = CursorRecordResult.Accept; // default; not used here
                     fixed (byte* keyPtr = key)
                     {
-                        if (NumUtils.HashSlot(keyPtr, key.Length) == slot && !Expired(ref value))
+                        if (HashSlotUtils.HashSlot(keyPtr, key.Length) == slot && !Expired(ref value))
                             keys.Add(key);
                     }
                     return true;
